requirements.txt
python-dotenv==1.0.0
pydantic==1.10.7
uvicorn==0.22.0
fastapi==0.95.1
websockets==10.1
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.3.0
pymongo==4.3.3